Fix clang warnings for linux builds (#6140)
https://github.com/youtube/cobalt/pull/6139 will actually remove the
clang warning suppression flags after this is submitted.

Bug: 374302273
Bug: 378934120

+++ media/starboard/progressive/avc_access_unit.cc
@@ -29,7 +29,7 @@ bool ReadBytes(uint64_t offset,
                size_t size,
                uint8_t* buffer,
                DataSourceReader* reader) {
-  if (reader->BlockingRead(offset, size, buffer) != size) {
+  if (reader->BlockingRead(offset, size, buffer) != static_cast<int>(size)) {
     LOG(ERROR) << "unable to download AU";
     return false;
   }
